+++ app/repository/SearchRepository.ts
@@ -1,6 +1,7 @@
 import { SingletonProto, AccessLevel, Inject } from '@eggjs/tegg';
 import { SearchAdapter } from '../common/typing';
 import { AuthorType, CnpmcorePatchInfo, PackageManifestType } from './PackageRepository';
+import { estypes } from '@elastic/elasticsearch';
 
 export type SearchJSONPickKey = '_rev' | 'name' | 'description' | 'keywords' | 'license' | 'maintainers' | 'dist-tags' | '_source_registry_name';
 
@@ -14,6 +15,10 @@ export type SearchMappingType = Pick<PackageManifestType, SearchJSONPickKey> & C
   author?: AuthorType | undefined;
   _npmUser?: {
     name: string;
+    email: string;
+  }
+  publisher?: {
+    username: string;
     email: string;
   }
 };
@@ -34,7 +39,7 @@ export class SearchRepository {
   private readonly searchAdapter: SearchAdapter;
 
 
-  async searchPackage(query) {
+  async searchPackage(query): Promise<estypes.SearchHitsMetadata<SearchManifestType>> {
     return await this.searchAdapter.search<SearchManifestType>(query);
   }